System.Process.Microlens.CommandSpec

Description

This module provides the associated optics and combinators for working with CommandSpec objects. CommandSpec consists of two cases: a Shell command, which is a command to execute naively in the shell, and a Raw command which is a command path together with its arguments.

CommandSpec has two cases, and therefore a Traversal into those two cases. There is also a convenient Traversal available for working with the arglist of a Raw command and combinators for working with arguments monoidally.

We provide classy variants for all useful traversals

Traversals

_RawCommand :: Traversal' CmdSpec (FilePath, [String])

A Traversal' into the RawCommand case of a CmdSpec

Examples:

>>> RawCommand "/bin/ls" ["-l"] ^? _RawCommand
Just ("/bin/ls",["-l"])

>>> RawCommand "/bin/ls" ["-l"] ^? _ShellCommand
Nothing

>>> RawCommand "/bin/ls" ["-l"] ^. _RawCommand . _1
"/bin/ls"

>>> RawCommand "/bin/ls" ["-l"] ^. _RawCommand . _2
["-l"]

_ShellCommand :: Traversal' CmdSpec String

A Traversal' into the ShellCommand case of a CmdSpec

Examples:

>>> ShellCommand "ls -l" ^? _ShellCommand
Just "ls -l"

>>> RawCommand "/bin/ls" ["-l"] ^? _ShellCommand
Nothing

arguments :: Traversal' CmdSpec [String]

Traversal' into the arguments of a command

Examples:

>>> RawCommand "/bin/ls" ["-l"] ^. arguments
["-l"]

Classy Traversals

class IsShell a where

Classy Traversal' into the shell command of a CmdSpec

Minimal complete definition

_Shell

Methods

_Shell :: Traversal' a String

class IsRaw a where

Classy Traversal' into the raw command of a CmdSpec

Minimal complete definition

_Raw

Methods

_Raw :: Traversal' a (FilePath, [String])

Combinators

arguing :: String -> CmdSpec -> CmdSpec

Append an argument to the argument list of a RawCommand

Examples:

>>> arguing "-h" $ RawCommand "/bin/ls" ["-l"]
RawCommand "/bin/ls" ["-l","-h"]

>>> arguing "-h" (RawCommand "/bin/ls" ["-l"]) ^. arguments
["-l","-h"]