System.Process.Microlens.CreateProcess

Description

This module provides the associated optics and combinators for working with CreateProcess objects.

Because CreateProcess was created before the `_` prefix record name convention, some record accessors don't have an apparently "good" name for their corresponding lens. Those that do not are post-fixed with `_`. Thankfully, there are 6 that meet the criteria: cmdspec_, env_, cwd_, stdin_, stdout_, and stderr_.

We provide classy variants of what we consider the significant portions of CreateProcess - namely, the std_in, std_out, and std_err entries.

Lenses

cmdspec_ :: Lens' CreateProcess CmdSpec

Lens into the cmdspec entry of the CreateProcess record

cwd_ :: Lens' CreateProcess (Maybe FilePath)

Lens into the cwd entry of the CreateProcess record

env_ :: Lens' CreateProcess (Maybe [(String, String)])

Lens into the env entry of the CreateProcess record

stdin_ :: Lens' CreateProcess StdStream

Lens into the std_in entry of the CreateProcess record

stdout_ :: Lens' CreateProcess StdStream

Lens into the std_out entry of the CreateProcess record

stderr_ :: Lens' CreateProcess StdStream

Lens into the std_err entry of the CreateProcess record

closefds :: Lens' CreateProcess Bool

Lens into the close_fds entry of the CreateProcess record

creategroup :: Lens' CreateProcess Bool

Lens into the create_group entry of the CreateProcess record

delegatectlc :: Lens' CreateProcess Bool

Lens into the delegate_ctlc entry of the CreateProcess record

newsession :: Lens' CreateProcess Bool

Lens into the new_session entry of the CreateProcess record

detachconsole :: Lens' CreateProcess Bool

Lens into the detach_console entry of the CreateProcess record

createnewconsole :: Lens' CreateProcess Bool

Lens into the create_new_console entry of the CreateProcess record

childgroup :: Lens' CreateProcess (Maybe CGid)

Lens into the child_group entry of the CreateProcess record

childuser :: Lens' CreateProcess (Maybe CUid)

Lens into the child_user entry of the CreateProcess record

useprocessjobs :: Lens' CreateProcess Bool

Lens into the use_process_jobs entry of the CreateProcess record

Classy Lenses

class HasStdin a where

Classy lens for types with a stdin

Minimal complete definition

_Stdin

Methods

_Stdin :: Lens' a StdStream

class HasStdout a where

Classy lens for types with a stdout

Minimal complete definition

_Stdout

Methods

_Stdout :: Lens' a StdStream

class HasStderr a where

Classy lens for types with a stderr

Minimal complete definition

_Stderr

Methods

_Stderr :: Lens' a StdStream