System.Process.Microlens.Optics

Description

Just the (classy) optics

Traversals

_ShellCommand :: Traversal' CmdSpec String

A Traversal' into the ShellCommand case of a CmdSpec

Examples:

>>> ShellCommand "ls -l" ^? _ShellCommand
Just "ls -l"

>>> RawCommand "/bin/ls" ["-l"] ^? _ShellCommand
Nothing

_RawCommand :: Traversal' CmdSpec (FilePath, [String])

A Traversal' into the RawCommand case of a CmdSpec

Examples:

>>> RawCommand "/bin/ls" ["-l"] ^? _RawCommand
Just ("/bin/ls",["-l"])

>>> RawCommand "/bin/ls" ["-l"] ^? _ShellCommand
Nothing

>>> RawCommand "/bin/ls" ["-l"] ^. _RawCommand . _1
"/bin/ls"

>>> RawCommand "/bin/ls" ["-l"] ^. _RawCommand . _2
["-l"]

_Inherit :: Traversal' StdStream StdStream

A Traversal' into the Inherit structure of a StdStream

_UseHandle :: Traversal' StdStream Handle

A Traversal' into the UseHandle structure's Handle for a StdStream

_CreatePipe :: Traversal' StdStream StdStream

A Traversal' into the CreatePipe structure of a StdStream

_NoStream :: Traversal' StdStream StdStream

A Traversal' into the NoStream structure of a StdStream

arguments :: Traversal' CmdSpec [String]

Traversal' into the arguments of a command

Examples:

>>> RawCommand "/bin/ls" ["-l"] ^. arguments
["-l"]

Lenses

cmdspec_ :: Lens' CreateProcess CmdSpec

Lens into the cmdspec entry of the CreateProcess record

cwd_ :: Lens' CreateProcess (Maybe FilePath)

Lens into the cwd entry of the CreateProcess record

env_ :: Lens' CreateProcess (Maybe [(String, String)])

Lens into the env entry of the CreateProcess record

stdin_ :: Lens' CreateProcess StdStream

Lens into the std_in entry of the CreateProcess record

stdout_ :: Lens' CreateProcess StdStream

Lens into the std_out entry of the CreateProcess record

stderr_ :: Lens' CreateProcess StdStream

Lens into the std_err entry of the CreateProcess record

closefds :: Lens' CreateProcess Bool

Lens into the close_fds entry of the CreateProcess record

creategroup :: Lens' CreateProcess Bool

Lens into the create_group entry of the CreateProcess record

delegatectlc :: Lens' CreateProcess Bool

Lens into the delegate_ctlc entry of the CreateProcess record

newsession :: Lens' CreateProcess Bool

Lens into the new_session entry of the CreateProcess record

detachconsole :: Lens' CreateProcess Bool

Lens into the detach_console entry of the CreateProcess record

createnewconsole :: Lens' CreateProcess Bool

Lens into the create_new_console entry of the CreateProcess record

childgroup :: Lens' CreateProcess (Maybe CGid)

Lens into the child_group entry of the CreateProcess record

childuser :: Lens' CreateProcess (Maybe CUid)

Lens into the child_user entry of the CreateProcess record

useprocessjobs :: Lens' CreateProcess Bool

Lens into the use_process_jobs entry of the CreateProcess record

hstdin :: Lens' ProcessHandler (Maybe Handle)

A lens into the stdin handle if requested

hstdout :: Lens' ProcessHandler (Maybe Handle)

A lens into the stdout handle if requested

hstderr :: Lens' ProcessHandler (Maybe Handle)

A lens into the stderr handle if requested

hhandle :: Lens' ProcessHandler ProcessHandle

A lens into the process handle

Classy Prisms

class IsRaw a where

Classy Traversal' into the raw command of a CmdSpec

Minimal complete definition

_Raw

Methods

_Raw :: Traversal' a (FilePath, [String])

class IsShell a where

Classy Traversal' into the shell command of a CmdSpec

Minimal complete definition

_Shell

Methods

_Shell :: Traversal' a String

class IsInherit a where

Class constraint proving a type has a prism into an Inherit structure. Any StdStream will have a prism into Inherit - it is just an overwrite to Inherit

Minimal complete definition

_Inherits

Methods

_Inherits :: Traversal' a StdStream

class IsUseHandle a where

Class constraint proving a type has a prism into a Handle via a UseHandle structure.

Minimal complete definition

_UsesHandle

Methods

_UsesHandle :: Traversal' a Handle

class IsCreatePipe a where

Class constraint proving a type has a prism into a Handle via a UseHandle structure. Any StdStream will have a prism into CreatePipe - it is just an overwrite to CreatePipe

Minimal complete definition

_CreatesPipe

Methods

_CreatesPipe :: Traversal' a StdStream

class IsNoStream a where

Class constraint proving a type has a prism into a Handle via a UseHandle structure. Any StdStream will have a prism into NoStream - it is just an overwrite to NoStream.

Minimal complete definition

_NoStreams

Methods

_NoStreams :: Traversal' a StdStream

Classy Lenses

class HasStdin a where

Classy lens for types with a stdin

Minimal complete definition

_Stdin

Methods

_Stdin :: Lens' a StdStream

class HasStdout a where

Classy lens for types with a stdout

Minimal complete definition

_Stdout

Methods

_Stdout :: Lens' a StdStream

class HasStderr a where

Classy lens for types with a stderr

Minimal complete definition

_Stderr

Methods

_Stderr :: Lens' a StdStream