System.Process.Microlens.ProcessHandler

Description

Convenient data type with associated optics + isos for working with the output of a createProcess call.

Types

data ProcessHandler

A convenient handler for the output of a createProcess call. This data containes 4 components:

a handle to stdin if it was requested

a handle to stdout if it was requested

a handle to stderr if it was requested

a process handle, containing a pid lock, information regarding ctcl-c delegation, and closed/open handle status info.

Constructors

ProcessHandler

Fields

_hstdin :: Maybe Handle

_hstdout :: Maybe Handle

_hstderr :: Maybe Handle

_hhandle :: ProcessHandle

Lenses

hstdin :: Lens' ProcessHandler (Maybe Handle)

A lens into the stdin handle if requested

hstdout :: Lens' ProcessHandler (Maybe Handle)

A lens into the stdout handle if requested

hstderr :: Lens' ProcessHandler (Maybe Handle)

A lens into the stderr handle if requested

hhandle :: Lens' ProcessHandler ProcessHandle

A lens into the process handle

Defaults

defaultCreateProcess :: CreateProcess

A default for a CreateProcess